Probability distributions

Modified versions of standard probability functions that can accommodate rvecs. If any of arguments to the functions are rvecs, then the functions return rvecs; otherwise they return ordinary R vectors,

Usage

dbeta_rvec(x, shape1, shape2, ncp = 0, log = FALSE)

pbeta_rvec(q, shape1, shape2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

qbeta_rvec(p, shape1, shape2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

rbeta_rvec(n, shape1, shape2, ncp = 0, n_draw = NULL)

dbinom_rvec(x, size, prob, log = FALSE)

pbinom_rvec(q, size, prob, lower.tail = TRUE, log.p = FALSE)

qbinom_rvec(p, size, prob, lower.tail = TRUE, log.p = FALSE)

rbinom_rvec(n, size, prob, n_draw = NULL)

dcauchy_rvec(x, location = 0, scale = 1, log = FALSE)

pcauchy_rvec(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qcauchy_rvec(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rcauchy_rvec(n, location = 0, scale = 1, n_draw = NULL)

dchisq_rvec(x, df, ncp = 0, log = FALSE)

pchisq_rvec(q, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

qchisq_rvec(p, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

rchisq_rvec(n, df, ncp = 0, n_draw = NULL)

dexp_rvec(x, rate = 1, log = FALSE)

pexp_rvec(q, rate = 1, lower.tail = TRUE, log.p = FALSE)

qexp_rvec(p, rate = 1, lower.tail = TRUE, log.p = FALSE)

rexp_rvec(n, rate = 1, n_draw = NULL)

df_rvec(x, df1, df2, ncp = 0, log = FALSE)

pf_rvec(q, df1, df2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

qf_rvec(p, df1, df2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

rf_rvec(n, df1, df2, ncp = 0, n_draw = NULL)

dgamma_rvec(x, shape, rate = 1, scale = 1/rate, log = FALSE)

pgamma_rvec(
  q,
  shape,
  rate = 1,
  scale = 1/rate,
  lower.tail = TRUE,
  log.p = FALSE
)

qgamma_rvec(
  p,
  shape,
  rate = 1,
  scale = 1/rate,
  lower.tail = TRUE,
  log.p = FALSE
)

rgamma_rvec(n, shape, rate = 1, scale = 1/rate, n_draw = NULL)

dgeom_rvec(x, prob, log = FALSE)

pgeom_rvec(q, prob, lower.tail = TRUE, log.p = FALSE)

qgeom_rvec(p, prob, lower.tail = TRUE, log.p = FALSE)

rgeom_rvec(n, prob, n_draw = NULL)

dhyper_rvec(x, m, n, k, log = FALSE)

phyper_rvec(q, m, n, k, lower.tail = TRUE, log.p = FALSE)

qhyper_rvec(p, m, n, k, lower.tail = TRUE, log.p = FALSE)

rhyper_rvec(nn, m, n, k, n_draw = NULL)

dlnorm_rvec(x, meanlog = 0, sdlog = 1, log = FALSE)

plnorm_rvec(q, meanlog = 0, sdlog = 1, lower.tail = TRUE, log.p = FALSE)

qlnorm_rvec(p, meanlog = 0, sdlog = 1, lower.tail = TRUE, log.p = FALSE)

rlnorm_rvec(n, meanlog = 0, sdlog = 1, n_draw = NULL)

dmultinom_rvec(x, size = NULL, prob, log = FALSE)

rmultinom_rvec(n, size, prob, n_draw = NULL)

dnbinom_rvec(x, size, prob, mu, log = FALSE)

pnbinom_rvec(q, size, prob, mu, lower.tail = TRUE, log.p = FALSE)

qnbinom_rvec(p, size, prob, mu, lower.tail = TRUE, log.p = FALSE)

rnbinom_rvec(n, size, prob, mu, n_draw = NULL)

dnorm_rvec(x, mean = 0, sd = 1, log = FALSE)

pnorm_rvec(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qnorm_rvec(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rnorm_rvec(n, mean = 0, sd = 1, n_draw = NULL)

dpois_rvec(x, lambda, log = FALSE)

ppois_rvec(q, lambda, lower.tail = TRUE, log.p = FALSE)

qpois_rvec(p, lambda, lower.tail = TRUE, log.p = FALSE)

rpois_rvec(n, lambda, n_draw = NULL)

dt_rvec(x, df, ncp = 0, log = FALSE)

pt_rvec(q, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

qt_rvec(p, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

rt_rvec(n, df, ncp = 0, n_draw = NULL)

dunif_rvec(x, min = 0, max = 1, log = FALSE)

punif_rvec(q, min = 0, max = 1, lower.tail = TRUE, log.p = FALSE)

qunif_rvec(p, min = 0, max = 1, lower.tail = TRUE, log.p = FALSE)

runif_rvec(n, min = 0, max = 1, n_draw = NULL)

dweibull_rvec(x, shape, scale = 1, log = FALSE)

pweibull_rvec(q, shape, scale = 1, lower.tail = TRUE, log.p = FALSE)

qweibull_rvec(p, shape, scale = 1, lower.tail = TRUE, log.p = FALSE)

rweibull_rvec(n, shape, scale = 1, n_draw = NULL)

Arguments

x

Quantiles. Can be rvec.

shape1, shape2

Parameters for beta distribution. Non-negative. See stats::dbeta(). Can be rvecs.

ncp

Non-centrality parameter. Default is 0. See stats::dbeta(), stats::dchisq(), stats::df(), stats::dt(). Cannot be rvec.

log, log.p

Whether to return log(p) rather than p. Default is FALSE. Can be rvec.

q

Quantiles. Can be rvec.

lower.tail

Whether to return \(P[X \le x]\), as opposed to \(P[X > x]\). Default is TRUE. Cannot be rvec.

p

Probabilities. Can be rvec.

n

In functions other than rhyper_rvec(), n is the length of random vector being created, and cannot be an rvec.

In rhyper_rvec(), n is the number of black balls in the urn, and can be an rvec. See stats::rhyper().

n_draw

Number of random draws, per observation, in random vector being created. Optional. Cannot be rvec.

size

Number of trials. See stats::dbinom(), stats::dmultinom(), stats::dnbinom(). Can be rvec.

prob

Probability of success in each trial. See stats::dgeom(), stats::dnbinom(). Can be rvec.

location

Parameter for Cauchy distribution. Default is 0. See stats::dcauchy(). Can be rvec.

scale

Parameter for Cauchy distribution. Default is 1. See stats::dcauchy(). Can be rvec.

df, df1, df2

Degrees of freedom. See stats::dchisq(), stats::df(), stats::dt(). Can be rvec.

rate

Rates. See stats::dexp(), stats::dgamma(). Can be rvec.

shape

Parameter for gamma distribution. See stats::dgamma(), stats::dweibull(). Can be rvec.

m

The number of white balls in the urn. See stats::dhyper(). Can be rvec.

k

Number of balls drawn from urn. See stats::dhyper(). Can be rvec.

nn

The length of the random vector being created, in a call to rhyper_rvec(). The equivalent of n in other random variate functions. See stats::rhyper(). Cannot be an rvec.

meanlog

Mean of distribution, on log scale. Default is 0. See stats::dlnorm(). Can be rvec.

sdlog

Standard deviation of distribution, on log scale. Default is 1. See stats::dlnorm(). Can be rvec.

mu

Mean for negative binomial distribution. See stats::dnbinom(). Can be rvec.

mean

Mean of distribution. Default is 0. See stats::dlnorm(). Can be rvec.

sd

Standard deviation. Default is 1. See stats::dnorm(). Can be rvec.

lambda

Vector of means. See stats::rpois() Can be rvec.

min, max

Lower and upper limits of the uniform distribution. Defaults are 0 and 1. See stats::dunif(). Can be rvec.

Value

If any of the arguments are rvecs, or if a value for n_draw is supplied, then an rvec; otherwise an ordinary R vector.

Creating an rvec by supplying a value for n_draw

The r*_rvec functions for generating random variates include one argument that is not present in base R functions: n_draw. If a value for n_draw is supplied, then the return value is always an rvec, even if none of the inputs are rvecs. This is a convenient way to create an rvec to use in a simulation. See below for an example.

Recycling

Unlike the base R distribution functions, the rvec functions use tidyverse vector recycling rules:

Vectors of length 1 are recycled

All other vectors must have the same size

These rules are more restrictive than base R rules, but are also more predictable. Base R style recycling can be done through explicit calls to base::rep(), base::rep_len(), and base::rep.int(), all of which have methods for rvecs.

Multinomial

The multinomial distribution is the only distribution described here that has a multivariate outcome. Base R and rvec multinomial distribution functions both behave differently from other distribution functions:

Base R	rvec

No pmultinom() or qmultinom()`	No pmultinom_rvec() or qmultinom_rvec()`

No recycling of arguments	No recycling of arguments

rmultinom() returns matrix	rmultinom_rvec() returns list if n > 1

See also

Examples

x_rv <- rvec(list(c(-0.8, 1.3),
                  c(-9.1, 8.7)))
mean_rv <- rvec(list(c(-1, 1),
                     c(-10, 10)))
sd_rv <- rvec(list(c(0.2, 20)))
x_rv
#> <rvec_dbl<2>[2]>
#> [1] -0.8,1.3 -9.1,8.7
mean_rv
#> <rvec_dbl<2>[2]>
#> [1] -1,1   -10,10
sd_rv
#> <rvec_dbl<2>[1]>
#> [1] 0.2,20

## densities: all arguments rvecs
dnorm_rvec(x = x_rv, mean = mean_rv, sd = sd_rv)
#> <rvec_dbl<2>[2]>
#> [1] 1.21,0.01994       0.00007992,0.01991

## densities: 'x' is ordinary vector
dnorm_rvec(x = c(0, 2), mean = mean_rv, sd = sd_rv)
#> <rvec_dbl<2>[2]>
#> [1] 0.000007434,0.01992 0,0.01841          

## ...which is equivalent to
c(dnorm_rvec(0, mean = mean_rv[1], sd = sd_rv),
  dnorm_rvec(2, mean = mean_rv[2], sd = sd_rv))
#> <rvec_dbl<2>[2]>
#> [1] 0.000007434,0.01992 0,0.01841          

## random variates: mean is rvec, sd is rvec
rnorm_rvec(n = 2, mean = mean_rv, sd = sd_rv)
#> <rvec_dbl<2>[2]>
#> [1] -1.025,14.48 -9.707,49.13

## random variates: mean is rvec, sd is scalar
rnorm_rvec(n = 2, mean = mean_rv, sd = 0.5)
#> <rvec_dbl<2>[2]>
#> [1] -1.135,0.8021 -10.62,10.05 

## create rvecs via the 'n_draw' argument,
## and use to created a simulated 'y'
mu <- rnorm_rvec(n = 3,
                 mean = 2,
                 sd = 0.5,
                 n_draw = 1000)
sigma <- rgamma_rvec(n = 3,
                     shape = 1,
                     scale = 0.5,
                     n_draw = 1000)
y <- rnorm_rvec(n = 3,
                mean = mu,
                sd = sigma)
y
#> <rvec_dbl<1000>[3]>
#> [1] 2 (0.11, 3.5) 2 (0.13, 3.6) 2 (0.29, 3.5)

## multinomial distribution
size <- rvec(list(10:12))
prob <- c(0.1, 0.4, 0.2, 0.3)
x <- rmultinom_rvec(n = 1, size = size, prob = prob)
x
#> <rvec_int<3>[4]>
#> [1] 1,0,0 1,5,6 2,0,2 6,6,4
sum(x)
#> <rvec_int<3>[1]>
#> [1] 10,11,12